Check in random notse files that were never checked in.
So I don't accidentally nuke them off my laptop again.

The main notes file was created back in 2017 while the core algorithm
reimplementation was still in flux, along with the romance dataset to
go with it. Some of the TODOs are genuinely TODOs, some others were
done quite a while ago.