Initial 'Rules' support!
This change adds Rules, which is an add-on to the base Correspondence
table. It functions by functionally essentially a 'staged synchronic
phronology module' pre-processing step to give the proto-forms which
are actually cross-compared with Correspondences.

* Add support to read and serialize to support new XML fields
specifying a rule table along with a correspondence table.
* Add support in the GTK gui to manipulate rules alongside
correspondences, as well as a supporting viewing the rule history in
the gui.
* Add lots of algorithm crap to actually support using the rules.
* Add two simple staged rules to the Slavic correspondence table
showing chronology.

Not done:

* Only the gui knows how to actually display cognate sets with forms
with rule histories properlyy, so other people won't see anything
useful right now. This is a huge FIXME.

* Probably tons of bugs lurking.