gtk: Create dialog picker for projects and settings.
This change creates a nice dialog picker for the projects and settings
to use. This way, people don't have to mess with the hairy command
line parser thing.

Also, in a push to rethink experiments, this change makes REGtk go
back to modifying the projects directory directly.

Update the README and howto accordingly.