gtk: Redesign menu item actions and prototype checkpoints feature.
* Change the save project action to a menu item based design and add
keyboard shortcuts for it and the zoom actions. This is more standard
GUI design than a save button associated with a specific
correspondence table widget. While this makes it more difficult to
save individual tables, the use case for purposely wanting to only
save individual parts of the input is rare. The concern is only
relevant for meso-language reconstructions anyway.

* Prototype the GUI model for the checkpoints feature. The actual
logic to create and load checkpoints is still unimplemented.