Make modern forms and attested lexicons comparable.
Via __eq__ and __hash__ methods. This gives us a lot of structural
equality testing and set operations for free while not slowing things
down in general too much. Use this to detect when to serialize out
lexicons as a smoke test, so we don't needlessly write out lexicons
every time we save.

This will be much more useful for computing diffs when comparing runs.