REgtk: Add a quick compare functionality.
Users can open this widget under View and the program keeps track of
the difference between the current run and the last run it made and
shows it in the diff view. The widget is made indepedently of the
window it displays in, not only so that we can reuse it in different
contexts, but also so that we can update the model in the background
even when the user has chosen to destroy the window. (so as to make it
appear as if the Window had actually been hidden)