RE: Read the entire ToC, not just the columns specified by upstream.
* Most importantly, this fixes the issue where serializing a partial
table overwrites the full table when using REgtk.
* Remove a bunch of confusing slots arguments, most of which were just
unused. We were conflating the languages to upstream, which is
specified in the settings object in an action tree, with the languages
present in a table of correspondence.
* This makes it obvious that the logic to create a parameter tree from
a settings file should not reference what is specified for
upstream. REgtk was doing this correctly.