Implement Panini's principle!
* This implementation tries to precompute as much stuff as possible
without blowing up the transition state space traveled. Two-sided
contexts which are more specific on both sides (and are used quite
extensively in the Romance data set) require actual logic to be added
to parsing itself in order to do context-matching on the side that
wasn't precomputed in the 1-step transition map. We have to keep track
of a bit more state in the recursive generation to do this.

* The rules for specificity are currently quite straightforward but do
cover almost all the cases we care about. Doing specificity based on
substringing is not done.

* Make the use of Panini toggleable, almost. It might be interesting
for the user in the future to diff with and without Panini to check
the effect or maybe find bugs in how they specify contexts.