Fix rule application to not generate multiple forms.
The problem was that, mainly at the end of a form, it could be
possible to apply the same rule multiple times if there are multiple
token lengths which yield the same substring when sliced.

Avoid this by generating based on index positions instead, which also
has the benefit of being faster as less strings need to be consed and
searching for rules can be stopped early based on length.